Move testTriggerRollingUpdateAfterOverrideBootstrap to AlternativeReconcileTriggersST
